+++ src/main/java/net/minecraft/client/renderer/entity/layers/DolphinCarryingItemLayer.java
@@ -40,9 +40,9 @@ public class DolphinCarryingItemLayer extends RenderLayer<Dolphin, DolphinModel<
         float var2 = -1.0F;
         float var3 = Mth.abs(param3.getXRot()) / 60.0F;
         if (param3.getXRot() < 0.0F) {
-            param0.translate(0.0, (double)(1.0F - var3 * 0.5F), (double)(-1.0F + var3 * 0.5F));
+            param0.translate(0.0F, 1.0F - var3 * 0.5F, -1.0F + var3 * 0.5F);
         } else {
-            param0.translate(0.0, (double)(1.0F + var3 * 0.8F), (double)(-1.0F + var3 * 0.2F));
+            param0.translate(0.0F, 1.0F + var3 * 0.8F, -1.0F + var3 * 0.2F);
         }
 
         ItemStack var4 = var0 ? param3.getMainHandItem() : param3.getOffhandItem();
